requirements.txt
--pre
pillow
pygetwindow
briefcase
detecto
toga
